+++ wgpu-core/src/validation/shader_io_deductions.rs
@@ -66,7 +66,7 @@ impl MaxFragmentShaderInputDeduction {
             BuiltIn::SubgroupSize => InterStageBuiltIn::SubgroupSize,
             BuiltIn::SubgroupInvocationId => InterStageBuiltIn::SubgroupInvocationId,
             BuiltIn::PointCoord => InterStageBuiltIn::PointCoord,
-            BuiltIn::Barycentric => InterStageBuiltIn::Barycentric,
+            BuiltIn::Barycentric { .. } => InterStageBuiltIn::Barycentric,
             BuiltIn::ViewIndex => InterStageBuiltIn::ViewIndex,
             BuiltIn::BaseInstance
             | BuiltIn::BaseVertex